service-mgmt-client/sm-client/requirements.txt
pbr>=0.5
anyjson>=0.3.3
argparse
httplib2
lxml>=2.3
PrettyTable>=0.6,<0.8
python-keystoneclient>=0.3.2
